+++ src/backend/core/models.py
@@ -5,11 +5,8 @@ Declare and configure the models for the impress core application
 
 import hashlib
 import smtplib
-import tempfile
-import textwrap
 import uuid
 from datetime import timedelta
-from io import BytesIO
 from logging import getLogger
 
 from django.conf import settings
@@ -21,19 +18,12 @@ from django.core.files.base import ContentFile
 from django.core.files.storage import default_storage
 from django.core.mail import send_mail
 from django.db import models
-from django.http import FileResponse
-from django.template.base import Template as DjangoTemplate
-from django.template.context import Context
 from django.template.loader import render_to_string
-from django.utils import html, timezone
+from django.utils import timezone
 from django.utils.functional import cached_property, lazy
 from django.utils.translation import get_language, override
 from django.utils.translation import gettext_lazy as _
 
-import frontmatter
-import markdown
-import pypandoc
-import weasyprint
 from botocore.exceptions import ClientError
 from timezone_field import TimeZoneField
 
@@ -793,107 +783,6 @@ class Template(BaseModel):
             "partial_update": can_update,
             "retrieve": can_get,
         }
-
-    def generate_pdf(self, body_html, metadata):
-        """
-        Generate and return a pdf document wrapped around the current template
-        """
-        document_html = weasyprint.HTML(
-            string=DjangoTemplate(self.code).render(
-                Context({"body": html.format_html(body_html), **metadata})
-            )
-        )
-        css = weasyprint.CSS(
-            string=self.css,
-            font_config=weasyprint.text.fonts.FontConfiguration(),
-        )
-
-        pdf_content = document_html.write_pdf(stylesheets=[css], zoom=1)
-        response = FileResponse(BytesIO(pdf_content), content_type="application/pdf")
-        response["Content-Disposition"] = f"attachment; filename={self.title}.pdf"
-
-        return response
-
-    def generate_word(self, body_html, metadata):
-        """
-        Generate and return a docx document wrapped around the current template
-        """
-        template_string = DjangoTemplate(self.code).render(
-            Context({"body": html.format_html(body_html), **metadata})
-        )
-
-        html_string = f"""
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <style>
-                {self.css}
-            </style>
-        </head>
-        <body>
-            {template_string}
-        </body>
-        </html>
-        """
-
-        reference_docx = "core/static/reference.docx"
-        output = BytesIO()
-
-        # Convert the HTML to a temporary docx file
-        with tempfile.NamedTemporaryFile(suffix=".docx", prefix="docx_") as tmp_file:
-            output_path = tmp_file.name
-
-            pypandoc.convert_text(
-                html_string,
-                "docx",
-                format="html",
-                outputfile=output_path,
-                extra_args=["--reference-doc", reference_docx],
-            )
-
-            # Create a BytesIO object to store the output of the temporary docx file
-            with open(output_path, "rb") as f:
-                output = BytesIO(f.read())
-
-        # Ensure the pointer is at the beginning
-        output.seek(0)
-
-        response = FileResponse(
-            output,
-            content_type="application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-        )
-        response["Content-Disposition"] = f"attachment; filename={self.title}.docx"
-
-        return response
-
-    def generate_document(self, body, body_type, export_format):
-        """
-        Generate and return a document for this template around the
-        body passed as argument.
-
-        2 types of body are accepted:
-        - HTML: body_type = "html"
-        - Markdown: body_type = "markdown"
-
-        2 types of documents can be generated:
-        - PDF: export_format = "pdf"
-        - Docx: export_format = "docx"
-        """
-        document = frontmatter.loads(body)
-        metadata = document.metadata
-        strip_body = document.content.strip()
-
-        if body_type == "html":
-            body_html = strip_body
-        else:
-            body_html = (
-                markdown.markdown(textwrap.dedent(strip_body)) if strip_body else ""
-            )
-
-        if export_format == "pdf":
-            return self.generate_pdf(body_html, metadata)
-
-        return self.generate_word(body_html, metadata)
 
 
 class TemplateAccess(BaseAccess):