+++ src/frontend/apps/impress/src/features/docs/doc-header/components/ModalExport.tsx
@@ -1,4 +1,12 @@
 import {
+  DOCXExporter,
+  docxDefaultSchemaMappings,
+} from '@blocknote/xl-docx-exporter';
+import {
+  PDFExporter,
+  pdfDefaultSchemaMappings,
+} from '@blocknote/xl-pdf-exporter';
+import {
   Button,
   Loader,
   Modal,
@@ -7,91 +15,116 @@ import {
   VariantType,
   useToastProvider,
 } from '@openfun/cunningham-react';
-import { useEffect, useMemo, useState } from 'react';
+import { pdf } from '@react-pdf/renderer';
+import { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { css } from 'styled-components';
 
 import { Box, Text } from '@/components';
 import { useEditorStore } from '@/features/docs/doc-editor';
 import { Doc } from '@/features/docs/doc-management';
 
-import { useExport } from '../api/useExport';
 import { TemplatesOrdering, useTemplates } from '../api/useTemplates';
-import { adaptBlockNoteHTML, downloadFile } from '../utils';
+import { downloadFile, exportResolveFileUrl } from '../utils';
 
-export enum DocDownloadFormat {
+enum DocDownloadFormat {
   PDF = 'pdf',
   DOCX = 'docx',
 }
 
-interface ModalPDFProps {
+interface ModalExportProps {
   onClose: () => void;
   doc: Doc;
 }
 
-export const ModalPDF = ({ onClose, doc }: ModalPDFProps) => {
+export const ModalExport = ({ onClose, doc }: ModalExportProps) => {
   const { t } = useTranslation();
   const { data: templates } = useTemplates({
     ordering: TemplatesOrdering.BY_CREATED_ON_DESC,
   });
   const { toast } = useToastProvider();
   const { editor } = useEditorStore();
-  const {
-    mutate: createExport,
-    data: documentGenerated,
-    isSuccess,
-    isPending,
-    error,
-  } = useExport();
-  const [templateIdSelected, setTemplateIdSelected] = useState<string>();
+  const [templateSelected, setTemplateSelected] = useState<string>('');
+  const [isExporting, setIsExporting] = useState(false);
   const [format, setFormat] = useState<DocDownloadFormat>(
     DocDownloadFormat.PDF,
   );
 
   const templateOptions = useMemo(() => {
-    if (!templates?.pages) {
-      return [];
-    }
-
-    const templateOptions = templates.pages
+    const templateOptions = (templates?.pages || [])
       .map((page) =>
         page.results.map((template) => ({
           label: template.title,
-          value: template.id,
+          value: template.code,
         })),
       )
       .flat();
 
-    if (templateOptions.length) {
-      setTemplateIdSelected(templateOptions[0].value);
-    }
+    templateOptions.unshift({
+      label: t('Empty template'),
+      value: '',
+    });
 
     return templateOptions;
-  }, [templates?.pages]);
+  }, [t, templates?.pages]);
 
-  useEffect(() => {
-    if (!error) {
+  async function onSubmit() {
+    if (!editor) {
+      toast(t('The export failed'), VariantType.ERROR);
       return;
     }
 
-    toast(error.message, VariantType.ERROR);
+    setIsExporting(true);
 
-    onClose();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [error, t]);
-
-  useEffect(() => {
-    if (!documentGenerated || !isSuccess) {
-      return;
-    }
-
-    // normalize title
     const title = doc.title
       .toLowerCase()
       .normalize('NFD')
       .replace(/[\u0300-\u036f]/g, '')
       .replace(/\s/g, '-');
 
-    downloadFile(documentGenerated, `${title}.${format}`);
+    const html = templateSelected;
+    let exportDocument = editor.document;
+    if (html) {
+      const blockTemplate = await editor.tryParseHTMLToBlocks(html);
+      exportDocument = [...blockTemplate, ...editor.document];
+    }
+
+    let blobExport: Blob;
+    if (format === DocDownloadFormat.PDF) {
+      const defaultExporter = new PDFExporter(
+        editor.schema,
+        pdfDefaultSchemaMappings,
+      );
+
+      const exporter = new PDFExporter(
+        editor.schema,
+        pdfDefaultSchemaMappings,
+        {
+          resolveFileUrl: async (url) =>
+            exportResolveFileUrl(url, defaultExporter.options.resolveFileUrl),
+        },
+      );
+      const pdfDocument = await exporter.toReactPDFDocument(exportDocument);
+      blobExport = await pdf(pdfDocument).toBlob();
+    } else {
+      const defaultExporter = new DOCXExporter(
+        editor.schema,
+        docxDefaultSchemaMappings,
+      );
+
+      const exporter = new DOCXExporter(
+        editor.schema,
+        docxDefaultSchemaMappings,
+        {
+          resolveFileUrl: async (url) =>
+            exportResolveFileUrl(url, defaultExporter.options.resolveFileUrl),
+        },
+      );
+
+      blobExport = await exporter.toBlob(exportDocument);
+    }
+
+    downloadFile(blobExport, `${title}.${format}`);
 
     toast(
       t('Your {{format}} was downloaded succesfully', {
@@ -100,29 +133,9 @@ export const ModalPDF = ({ onClose, doc }: ModalPDFProps) => {
       VariantType.SUCCESS,
     );
 
+    setIsExporting(false);
+
     onClose();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [documentGenerated, isSuccess, t]);
-
-  async function onSubmit() {
-    if (!templateIdSelected || !format) {
-      return;
-    }
-
-    if (!editor) {
-      toast(t('No editor found'), VariantType.ERROR);
-      return;
-    }
-
-    let body = await editor.blocksToFullHTML(editor.document);
-    body = adaptBlockNoteHTML(body);
-
-    createExport({
-      templateId: templateIdSelected,
-      body,
-      body_type: 'html',
-      format,
-    });
   }
 
   return (
@@ -138,6 +151,7 @@ export const ModalPDF = ({ onClose, doc }: ModalPDFProps) => {
             color="secondary"
             fullWidth
             onClick={() => onClose()}
+            disabled={isExporting}
           >
             {t('Cancel')}
           </Button>
@@ -146,7 +160,7 @@ export const ModalPDF = ({ onClose, doc }: ModalPDFProps) => {
             color="primary"
             fullWidth
             onClick={() => void onSubmit()}
-            disabled={isPending || !templateIdSelected}
+            disabled={isExporting}
           >
             {t('Download')}
           </Button>
@@ -173,9 +187,9 @@ export const ModalPDF = ({ onClose, doc }: ModalPDFProps) => {
           clearable={false}
           label={t('Template')}
           options={templateOptions}
-          value={templateIdSelected}
+          value={templateSelected}
           onChange={(options) =>
-            setTemplateIdSelected(options.target.value as string)
+            setTemplateSelected(options.target.value as string)
           }
         />
         <Select
@@ -192,8 +206,17 @@ export const ModalPDF = ({ onClose, doc }: ModalPDFProps) => {
           }
         />
 
-        {isPending && (
-          <Box $align="center" $margin={{ top: 'big' }}>
+        {isExporting && (
+          <Box
+            $align="center"
+            $margin={{ top: 'big' }}
+            $css={css`
+              position: absolute;
+              left: 50%;
+              top: 50%;
+              transform: translate(-50%, -100%);
+            `}
+          >
             <Loader />
           </Box>
         )}