+++ src/frontend/apps/impress/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './Crisp';
+export * from './Posthog';